Changed controls
Controls changed B to X
Reversed catapult motor
When button is release intake now starts
Fixed float vs int bugs for flipper and descorer
Fixed boolean subtraction for descorer

+++ include/descorer.h
@@ -1,3 +1,3 @@
 #pragma once
 
-void setDescorerManual(int speed);
+void setDescorerManual(float speed);
